force ssh checkouts
remove older scripts

+++ bin/pull-all.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 dir="${0%[/\\]*}"
 pushd $dir
-npm install bigstraw
+git pull
+if [ ! -e node_modules/bigstraw ]; then
+  npm install bigstraw
+fi
 popd
-node $dir/node_modules/bigstraw/index.js $dir/../repo-configs/*.json $@
+node $dir/node_modules/bigstraw/index.js -s $dir/../repo-configs/*.json $@
